add middleware

+++ routes/api.php
@@ -10,31 +10,20 @@ Route::middleware(['auth:sanctum'])->group(function () {
     Route::get('/user', function (Request $request) {
         return $request->user();
     });
+});
+
     /*
     |------------------------------------------------------------------
     | -  User Routes
     |------------------------------------------------------------------
      */
     Route::prefix('users')->group(function () {
-        Route::get('/', [UserController::class, 'index']);
-        Route::get('/{user}', [UserController::class, 'show']);
-        Route::post('/', [UserController::class, 'store']);
-        Route::put('/{user}', [UserController::class, 'update']);
-        Route::delete('/{user}', [UserController::class, 'destroy']);
+        Route::get('/', [UserController::class, 'index'])->middleware('permission:view-users');
+        Route::get('/{user}', [UserController::class, 'show'])->middleware('permission:view-users');
+        Route::post('/', [UserController::class, 'store'])->middleware('permission:create-user');
+        Route::put('/{user}', [UserController::class, 'update'])->middleware('permission:edit-user');
+        Route::delete('/{user}', [UserController::class, 'destroy'])->middleware('permission:delete-user');
     });
-    /*
-    |------------------------------------------------------------------
-    |  -  Role Routes
-    |------------------------------------------------------------------
-     */
-    // Route::prefix('roles')->group(function () {
-    //     Route::get('/', [RoleController::class, 'index'])->name('roles.index');
-    //     Route::get('/{role}', [RoleController::class, 'show'])->name('roles.show');
-    //     Route::post('/', [RoleController::class, 'store'])->name('roles.store');
-    //     Route::put('/{role}', [RoleController::class, 'update'])->name('roles.update');
-    //     Route::delete('/{role}', [RoleController::class, 'destroy'])->name('roles.destroy');
-    // });
-});
 
 require __DIR__.'/site.php';
 require __DIR__.'/auth.php';
